refactor(attributes): moved the Attributes class into the `internal` package
