+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/AttributesJson.kt
@@ -3,7 +3,6 @@
 package com.bugsnag.android.performance.internal
 
 import android.util.JsonWriter
-import com.bugsnag.android.performance.Attributes
 import kotlin.collections.forEach
 
 internal fun JsonWriter.value(attributes: Attributes): JsonWriter {
